put mapSigningRequestToState in its own file and class
