+++ lib/v2/blocs/deeplink/viewmodels/deeplink_bloc.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:firebase_dynamic_links/firebase_dynamic_links.dart';
 import 'package:seeds/v2/blocs/deeplink/mappers/deep_link_state_mapper.dart';
+import 'package:seeds/v2/blocs/deeplink/mappers/eosio_signing_request_state_mapper.dart';
 import 'package:seeds/v2/blocs/deeplink/usecase/get_initial_deep_link_use_case.dart';
 import 'package:seeds/v2/blocs/deeplink/usecase/get_signing_request_use_case.dart';
 import 'package:seeds/v2/blocs/deeplink/viewmodels/deeplink_event.dart';
@@ -35,7 +36,7 @@ class DeeplinkBloc extends Bloc<DeeplinkEvent, DeeplinkState> {
       yield DeeplinkState.initial();
     } else if (event is HandleIncomingSigningRequest) {
       final result = await GetSigningRequestUseCase().run(event.link);
-      yield DeepLinkStateMapper().mapSigningRequestToState(state, result);
+      yield EosioSigningRequestStateMapper().mapSigningRequestToState(state, result);
     }
   }
